requirements.txt
asgiref==3.9.1
Brotli==1.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.9.0
django-filter==25.1
django-rest-passwordreset==1.5.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
h11==0.16.0
idna==3.10
packaging==25.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
urllib3==2.5.0
uvicorn==0.37.0
whitenoise==6.11.0
